+++ bignetcoin-server/src/main/java/com/bignetcoin/server/transaction/FreeStandingTransactionOutput.java
@@ -40,7 +40,7 @@ public class FreeStandingTransactionOutput extends TransactionOutput {
      * @return The depth.
      */
     @Override
-    public int getParentTransactionDepthInBlocks() {
+    public long getParentTransactionDepthInBlocks() {
         return chainHeight - output.getHeight() + 1;
     }
